Modifie le temps d'une simulation pour qu'il soit de type Double

+++ src/main/scala/sim/Simulator.scala
@@ -12,7 +12,7 @@ import model.{Solution, Reaction}
 class Simulator(
   initialSolution: Solution,
   reactions: Set[Reaction],
-  maxTime: Int
+  maxTime: Double
 ) {
 
   import scala.util.Random
